refactor windows cgo
